fix: add all changes to create page
